# Conflicts:
#	project.properties

+++ org.jenetics/src/main/java/org/jenetics/IntegerChromosome.java
@@ -57,7 +57,13 @@ public class IntegerChromosome
 {
 	private static final long serialVersionUID = 1L;
 
-
+	/**
+	 * Create a new chromosome from the given genes array.
+	 *
+	 * @param genes the genes of the new chromosome.
+	 * @throws IllegalArgumentException if the gene sequence is empty
+	 * @throws NullPointerException if the {@code genes} are {@code null}.
+	 */
 	protected IntegerChromosome(final ISeq<IntegerGene> genes) {
 		super(genes);
 	}
@@ -69,8 +75,14 @@ public class IntegerChromosome
 	 * @param max the max value of the {@link IntegerGene}s (inclusively).
 	 * @param length the length of the chromosome.
 	 * @throws NullPointerException if one of the arguments is {@code null}.
+	 * @throws IllegalArgumentException if the {@code length} is smaller than
+	 *         one.
 	 */
-	public IntegerChromosome(final Integer min, final Integer max, final int length) {
+	public IntegerChromosome(
+		final Integer min,
+		final Integer max,
+		final int length
+	) {
 		this(IntegerGene.seq(min, max, length));
 		_valid = true;
 	}
@@ -142,6 +154,7 @@ public class IntegerChromosome
 	 * @param max the max value of the {@link IntegerGene}s (inclusively).
 	 * @param length the length of the chromosome.
 	 * @return a new random {@code IntegerChromosome}
+	 * @throws IllegalArgumentException if the length is smaller than one
 	 */
 	public static IntegerChromosome of(
 		final int min,
@@ -160,6 +173,7 @@ public class IntegerChromosome
 	 * @param length the length of the chromosome.
 	 * @return a new random {@code IntegerChromosome}
 	 * @throws NullPointerException if the given {@code range} is {@code null}
+	 * @throws IllegalArgumentException if the length is smaller than one
 	 */
 	public static IntegerChromosome of(final IntRange range, final int length) {
 		return new IntegerChromosome(range.getMin(), range.getMax(), length);
@@ -219,11 +233,11 @@ public class IntegerChromosome
 		out.defaultWriteObject();
 
 		out.writeInt(length());
-		out.writeInt(_min.intValue());
-		out.writeInt(_max.intValue());
+		out.writeInt(_min);
+		out.writeInt(_max);
 
 		for (IntegerGene gene : _genes) {
-			out.writeInt(gene.getAllele().intValue());
+			out.writeInt(gene.getAllele());
 		}
 	}
